Add automatical migation guard
