+++ src/Hedwig/appsettings.json
@@ -23,5 +23,6 @@
 	},
 	"AllowedHosts": "*",
 	"WingedKeysUri": "https://localhost:5050",
-	"EnvironmentName": "Development"
+	"EnvironmentName": "Development",
+	"AutomaticallyApplyMigrations": "false"
 }
